did alot off stuff
