+++ SpaceGame/Assets/Prefabs/RotatorScreen 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 853366cd3963a8a41932af4ad644cf85
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
