+++ SpaceGame/Assets/Prefabs/Spline (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bf99b503a7d50e4469445318c6c92808
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
