+++ SpaceGame/Assets/Prefabs/generated by SplineExtrusion.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f7d03a83cb1d3814ba41399df83879ee
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
