+++ SpaceGame/Assets/Scenes/SampleScene.unity
@@ -25975,7 +25975,7 @@ MonoBehaviour:
   m_MaxCount: 3
   m_targetRenderer: {fileID: 2087553231}
   m_NormalColor: {r: 0, g: 0.92156863, b: 0.9019608, a: 1}
-  m_FullColor: {r: 0.509434, g: 0, b: 0.009582835, a: 1}
+  m_FullColor: {r: 1, g: 0, b: 0.015384674, a: 1}
 --- !u!4 &2110639914 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 8684410594036502101, guid: f393324e9b96414469d9089404af39aa,
