+++ SpaceGame/Assets/Scripts/UpdateBuoyText.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 39d5cb8dfdf35694bbd811ccc86ec101
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
