+++ SpaceGame/Assets/VFX/Shield Shader (Sucess)/Shield Shader.mat
@@ -7,7 +7,7 @@ Material:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: Shader Graphs_New Shader Graph 1
+  m_Name: Shield Shader
   m_Shader: {fileID: -6465566751694194690, guid: a59446d22d647df41910aa271f15d86b,
     type: 3}
   m_ShaderKeywords:
